+++ decorator.rb
@@ -11,18 +11,5 @@ class Decorator < Nameable
   end
 end
 
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
 
-class TrimmerDecorator < Decorator
-  def correct_name
-    if @nameable.correct_name.size >= 11
-      @nameable.correct_name.slice(0, 10)
-    else
-      @nameable.correct_name
-    end
-  end
-end
+
